requirements.txt
torch>=2.0.0
transformers>=4.55.0
Pillow>=9.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
PyMuPDF>=1.21.0  # For PDF processing
requests>=2.28.0
